+ htable.h; dlist funcs added; misc

+++ src/list.h
@@ -1,12 +1,14 @@
 #ifndef __LIST_H_
 #define __LIST_H_
 
+#include <stdlib.h>
+
 typedef struct node {
     struct node *next;
 } node;
 
 typedef struct list {
-    int count;
+    size_t count;
     node *first;
 } list;
 
@@ -16,7 +18,7 @@ typedef struct cs_node {
 } cs_node;
 
 typedef struct cs_list {
-    int count;
+    size_t count;
     cs_node *first;
 } cs_list;
